Ref: Extract Lightweight TUI package w keystroke/layout/theme/widget
